Systematic decode audit: fix INT4->i64 and FLOAT4->f64 decode bugs
The real-DB integration suite is dormant on this host (the `pgmcp` role lacks
CREATEDB, so require_test_db! early-returns as "passed"), which hides runtime
sqlx decode-type mismatches. Add `scripts/decode-audit.sh` — a read-only audit
that runs every read-only tool against the live DB and checks cron_run_history,
flagging `mismatched types` / `decoding column` errors.

The audit surfaced 3 bugs in 2 classes the earlier NUMERIC sweep didn't cover;
fixed at the QUERY level so every tool/cron sharing the query is fixed:

- INT4 -> i64: `code_topics.id` and `chunk_topic_assignments.topic_id` are
  `integer`, but several FromRow structs decode `topic_id` as `i64`/`Option<i64>`
  (sqlx rejects INT4->i64). Cast `::bigint` at the select in:
  - db/queries/chunks.rs (get_chunk_topic_summaries -> chunk_clusters)
  - db/queries/similarity.rs (find_pattern_abstraction_pairs ->
    pattern_abstraction_candidates)
  - db/queries/metrics.rs (get_god_file_chunks_with_topics ->
    recommend_module_split)

- FLOAT4 -> f64: `symbol_references.resolution_confidence` is `real`, decoded as
  `Option<f64>`. Cast `::float8` in db/queries/graph.rs (list_call_edges_for_project
  -> recursive_clusters + the call_graph cron).

All four fixes verified against the live DB (tools now return rows with no decode
error). A full multi-method sweep (empirical read-only tools + cron_run_history +
a static column-type-vs-Rust-type cross-check) found no further mismatches: the
codebase is otherwise decode-clean (topics.rs correctly uses i32 for the INT4
topic_id; quality GPAs use f32 for their `real` columns; graph weights are
`double precision`). NOT migration regressions — sqlx 0.8 rejected these too.

+++ src/db/queries/chunks.rs
@@ -509,7 +509,9 @@ pub async fn get_chunk_topic_summaries(
         return Ok(Vec::new());
     }
     sqlx::query_as::<_, ChunkTopicSummaryRow>(
-        "SELECT cta.chunk_id, ct.id AS topic_id, ct.label, ct.keywords, cta.membership_score
+        // `code_topics.id` is `integer` (INT4); cast to BIGINT so it decodes into
+        // the `topic_id: i64` field (sqlx rejects INT4->i64).
+        "SELECT cta.chunk_id, ct.id::bigint AS topic_id, ct.label, ct.keywords, cta.membership_score
          FROM chunk_topic_assignments cta
          JOIN code_topics ct ON ct.id = cta.topic_id
          WHERE cta.chunk_id = ANY($1)
